PI Assignment
[4]
Class Library Product


create new class library project
Name as ShoppingCartLib

Rename class as Product

----

create three full property
ProductName : string
Price : decimal
Quantity : int

---
create constructor - no para,

create param constructor
	take all 3 values via constructor
	assign to respective properties

---
Create method as Get Productcost
methodname : ProductCost
Param : no param
price * quantity

---
override ToString method - print all 3 values

---
build you project

==============================
create console application in same solution

Name as AmazonApp

Add library reference

add using statement

Create object of Product call and test the code







+++++++++++++++++++++++++++++++++++++++++++++++++++


5.	Print Armstrong number between 1 to 1000

An Armstrong number of three digits is an integer such that the sum of the cubes of its digits is equal to the number itself. For example, 371 is an Armstrong number since 3**3 + 7**3 + 1**3 = 371


6.	Print Neon number between 1 to 1000
A neon number is a number where the sum of digits of square of the number is equal to the number
For example if the input number is 9, its square is 9*9 = 81 and sum of the digits is 9. i.e. 9 is a neon number


Array in C#

7.	Implement Binary search 
Binary search is a search algorithm that finds the position of a target value within a sorted array. Binary search compares the target value to the middle element of the array; if they are unequal, the half in which the target cannot lie is eliminated and the search continues on the remaining half until it is successful. If the search ends with the remaining half being empty, the target is not in the array.


8.	Implement Bubble Sort
Bubble sort is a simple sorting algorithm that repeatedly steps through the list to be sorted, compares each pair of adjacent items and swaps them if they are in the wrong order. The pass through the list is repeated until no swaps are needed, which indicates that the list is sorted